+++ lib/IDatabaseConnection.d.ts
@@ -2,6 +2,7 @@
 import { Readable } from 'stream';
 import { Query } from './Query';
 export interface IDatabaseConnection {
+    setInstantiationStack(stack: string): void;
     getInstantiationStack(): string;
     getAPI(): any;
     isReadOnly(): boolean;